+++ src/VBox/Main/src-server/generic/NetIf-generic.cpp
@@ -112,7 +112,7 @@ int NetIfAdpCtlOut(const char * pcszName, const char * pcszCmd, char *pszBuffer,
     FILE *fp = popen(szAdpCtl, "r");
     if (fp)
     {
-        if (fgets(pszBuffer, cBufSize, fp))
+        if (fgets(pszBuffer, (int)cBufSize, fp))
         {
             if (!strncmp(VBOXNETADPCTL_NAME ":", pszBuffer, sizeof(VBOXNETADPCTL_NAME)))
             {
